risteyslaskenta_package/__init__.py
import os

from qgis.gui import QgisInterface

from risteyslaskenta_package.qgis_plugin_tools.infrastructure.debugging import (  # noqa F401
    setup_debugpy,
    setup_ptvsd,
    setup_pydevd,
)

debugger = os.environ.get("QGIS_PLUGIN_USE_DEBUGGER", "").lower()
if debugger in {"debugpy", "ptvsd", "pydevd"}:
    locals()["setup_" + debugger]()


def classFactory(iface: QgisInterface):  # noqa N802
    from risteyslaskenta_package.plugin import Plugin

    return Plugin()
