requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements-dev.in
#
astor==0.8.1
    # via flake8-qgis
attrs==22.1.0
    # via
    #   flake8-annotations
    #   flake8-bugbear
    #   pytest
black==22.10.0
    # via -r requirements-dev.in
build==0.9.0
    # via pip-tools
cfgv==3.3.1
    # via pre-commit
click==8.1.3
    # via
    #   black
    #   pip-tools
coverage[toml]==6.5.0
    # via pytest-cov
distlib==0.3.6
    # via virtualenv
exceptiongroup==1.0.4
    # via pytest
filelock==3.8.2
    # via virtualenv
flake8==6.0.0
    # via
    #   -r requirements-dev.in
    #   flake8-annotations
    #   flake8-bugbear
    #   flake8-qgis
    #   pep8-naming
flake8-annotations==2.9.1
    # via -r requirements-dev.in
flake8-bugbear==22.12.6
    # via -r requirements-dev.in
flake8-qgis==1.0.0
    # via -r requirements-dev.in
identify==2.5.9
    # via pre-commit
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via -r requirements-dev.in
mccabe==0.7.0
    # via flake8
mypy==0.991
    # via -r requirements-dev.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
nodeenv==1.7.0
    # via pre-commit
packaging==22.0
    # via
    #   build
    #   pytest
pathspec==0.10.2
    # via black
pep517==0.13.0
    # via build
pep8-naming==0.13.2
    # via -r requirements-dev.in
pip-tools==6.11.0
    # via -r requirements-dev.in
platformdirs==2.6.0
    # via
    #   black
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.20.0
    # via -r requirements-dev.in
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pyqt5-stubs==5.15.6.0
    # via -r requirements-dev.in
pytest==7.2.0
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-qgis
pytest-cov==4.0.0
    # via -r requirements-dev.in
pytest-qgis==1.3.2
    # via -r requirements-dev.in
pyyaml==6.0
    # via pre-commit
toml==0.10.2
    # via pre-commit
tomli==2.0.1
    # via
    #   black
    #   build
    #   coverage
    #   mypy
    #   pytest
typing-extensions==4.4.0
    # via mypy
virtualenv==20.17.1
    # via pre-commit
wheel==0.38.4
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
